requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
appnope
asgiref
asttokens
attrs
bcrypt
beautifulsoup4
blinker
cachetools
certifi
cffi
charset-normalizer
click
comm
contourpy
cryptography
cycler
dataclasses-json
debugpy
decorator
distro
Django
djangorestframework
dnspython
docx2txt
ecdsa
email_validator
executing
fastapi
filelock
filetype
Flask
Flask-Bcrypt
flask-cors
Flask-JWT-Extended
Flask-Login
Flask-Mail
Flask-MySQL
Flask-SQLAlchemy
fonttools
frozenlist
fsspec
fuzzywuzzy
gitdb
GitPython
google-generativeai
google-ai-generativelanguage
google-api-core
google-auth
googleapis-common-protos
grpcio
grpcio-status
h11
httpcore
httpx
httpx-sse
huggingface-hub
idna
ipykernel
ipython
ipython_pygments_lexers
itsdangerous
jedi
Jinja2
jiter
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
kiwisolver
protobuf>=4.25.3,<7
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
python-jose
langsmith
lxml
markdown-it-py
markdownify
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mdurl
mpmath
multidict
mypy-extensions
mysql-connector-python
mysqlclient
narwhals
nest-asyncio
networkx
numpy
openai
orjson
packaging
pandas
parso
passlib
pexpect
pillow
platformdirs
plotly
prompt_toolkit
propcache
proto-plus
psutil
ptyprocess
pure_eval
pyarrow
pyasn1
pyasn1-modules
pycparser
pydantic
pydantic-settings
pydantic_core
pydeck
Pygments
PyJWT
pymongo
PyMySQL
pyparsing
PyPDF2
python-dateutil
python-docx
python-dotenv
python-jobspy
python-multipart
pytz
PyYAML
pyzmq
referencing
regex
requests
requests-toolbelt
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
seaborn
setuptools
six
smmap
sniffio
soupsieve
SQLAlchemy
sqlparse
stack-data
starlette
streamlit
sympy
tenacity
threadpoolctl
tls-client
tokenizers
toml
torch
tornado
tqdm
traitlets
transformers
typing-inspect
typing_extensions
tzdata
urllib3
uvicorn
wcwidth
Werkzeug
yarl
gunicorn
zstandard
psycopg2-binary
gevent
